refactor: Move enter action into its own class.
